pyproject.toml
[tool.poetry]
name = "ta-learn"
version = "0.1.0"
description = ""
authors = ["scriptmoney <hello@script.money>"]
readme = "README.md"
packages = [{include = "ta_learn"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.1"
ccxt = "^2.1.93"
openai = "^0.25.0"
python-dotenv = "^0.21.0"
plotly = "^5.11.0"
apscheduler = "^3.9.1.post1"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


#设置poetry包管理工具的自定义pypi镜像源配置
[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
default = true

# Python-style entrypoints and scripts are easily expressed
[tool.poetry.scripts]
sm = "sm:main"